+++ src/index.ts
@@ -2,7 +2,9 @@ import { Flow } from './components/flow';
 export { Flow } from './components/flow';
 export default Flow;
 
-export * from './api/types/core';
-export * from './api/types/extended';
+export type * from './api/types/core';
+export type * from './api/types/extended';
 
 export * as FlowActions from './components/action';
+export type { FlowQuery } from './components/query';
+export type { FlowResponse, FlowActionResponse } from './components/response';